+++ adm/app/adms/Views/colors/listColors.php
@@ -16,9 +16,9 @@ if (isset($this->data['form'])) {
             <span class="title-content">Listar Cores</span>
             <div class="top-list-right">
                 <?php
-              
+                if( $this->data['button']['add_colors']){
                     echo "<a href='" . URLADM . "add-colors/index' class='btn-success'>Cadastrar</a>";
-                        
+                }
                 ?>
             </div>
         </div>
@@ -87,14 +87,15 @@ if (isset($this->data['form'])) {
                                 <button onclick="actionDropdown(<?php echo $id; ?>)" class="dropdown-btn-action">Ações</button>
                                 <div id="actionDropdown<?php echo $id; ?>" class="dropdown-action-item">
                                     <?php                                    
-                              
+                               if( $this->data['button']['view_colors']){
                                         echo "<a href='" . URLADM . "view-colors/index/$id'>Visualizar</a>";
-                          
+                               }
+                               if( $this->data['button']['edit_colors']){
                                         echo "<a href='" . URLADM . "edit-colors/index/$id'>Editar</a>";
-                             
-                                    
+                               }
+                               if( $this->data['button']['delete_colors']){
                                         echo "<a href='" . URLADM . "delete-colors/index/$id' onclick='return confirm(\"Tem certeza que deseja excluir este registro?\")'>Apagar</a>";
-                           
+                               }
                                     ?>
                                 </div>
                             </div>
