+++ adm/app/adms/Views/groupsPages/viewGroupsPages.php
@@ -11,10 +11,16 @@ if(!defined('C8L6K7E')){
             <span class="title-content">Detalhes do Grupo de Página</span>
             <div class="top-list-right">
                 <?php
+                  if( $this->data['button']['list_groups_pages']){
                 echo "<a href='" . URLADM . "list-groups-pages/index' class='btn-info'>Listar</a> ";
+                  }
                 if (!empty($this->data['viewGroupsPages'])) {
+                    if( $this->data['button']['edit_groups_pages']){
                     echo "<a href='" . URLADM . "edit-groups-pages/index/" . $this->data['viewGroupsPages'][0]['id'] . "' class='btn-warning'>Editar</a> ";
+                    }
+                    if( $this->data['button']['delete_groups_pages']){
                     echo "<a href='" . URLADM . "delete-groups-pages/index/" . $this->data['viewGroupsPages'][0]['id'] . "' onclick='return confirm(\"Tem certeza que deseja excluir este registro?\")' class='btn-danger'>Apagar</a> ";
+                    }
                 }
                 ?>
             </div>
