+++ adm/app/adms/Views/pages/listPages.php
@@ -16,8 +16,12 @@ if (isset($this->data['form'])) {
             <span class="title-content">Listar Páginas</span>
             <div class="top-list-right">
                 <?php
+                 if( $this->data['button']['add_pages']){
                 echo "<a href='" . URLADM . "add-pages/index' class='btn-success'>Cadastrar</a> ";
+                 }
+                 if( $this->data['button']['sync_pages_levels']){
                 echo "<a href='" . URLADM . "sync-pages-levels/index' class='btn-warning'>Sincronizar</a>";
+                 }
                 ?>
             </div>
         </div>
@@ -58,9 +62,15 @@ if (isset($this->data['form'])) {
                                 <button onclick="actionDropdown(<?php echo $id; ?>)" class="dropdown-btn-action">Ações</button>
                                 <div id="actionDropdown<?php echo $id; ?>" class="dropdown-action-item">
                                     <?php
+                                     if( $this->data['button']['view_pages']){
                                     echo "<a href='" . URLADM . "view-pages/index/$id'>Visualizar</a>";
+                                     }
+                                     if( $this->data['button']['edit_pages']){
                                     echo "<a href='" . URLADM . "edit-pages/index/$id'>Editar</a>";
+                                     }
+                                     if( $this->data['button']['delete_pages']){
                                     echo "<a href='" . URLADM . "delete-pages/index/$id' onclick='return confirm(\"Tem certeza que deseja excluir este registro?\")'>Apagar</a>";
+                                     }
                                     ?>
                                 </div>
                             </div>
