+++ adm/app/adms/Views/sitsPages/listSitPages.php
@@ -15,7 +15,9 @@ if (isset($this->data['form'])) {
             <span class="title-content">Listar Situação da Página</span>
             <div class="top-list-right">
                 <?php
+                 if( $this->data['button']['add_sits_pages']){
                 echo "<a href='" . URLADM . "add-sits-pages/index' class='btn-success'>Cadastrar</a>";
+                 }
                 ?>
             </div>
 
@@ -54,9 +56,15 @@ if (isset($this->data['form'])) {
                                 <div id="actionDropdown<?php echo $id ?>" class="dropdown-action-item">
 
                                     <?php
+                                     if( $this->data['button']['view_sits_pages']){
                                     echo "<a href='" . URLADM . "view-sits-pages/index/$id'>Visualizar</a>";
+                                     }
+                                     if( $this->data['button']['edit_sits_pages']){
                                     echo "<a href='" . URLADM . "edit-sits-pages/index/$id'>Editar</a>";
+                                     }
+                                     if( $this->data['button']['delete_sits_pages']){
                                     echo "<a href='" . URLADM . "delete-sits-pages/index/$id' onclick='return confirm(\"Tem certeza que deseja excluir este registro?\")'>Apagar</a>";
+                                     }
                                     ?>
                                 </div>
                             </div>
