+++ adm/app/adms/Views/sitsPages/viewSitPages.php
@@ -11,10 +11,16 @@ if(!defined('C8L6K7E')){
             <span class="title-content">Detalhes da Situação de Página</span>
             <div class="top-list-right">
                 <?php
+                 if( $this->data['button']['list_sits_pages']){
                 echo "<a href='" . URLADM . "list-sits-pages/index' class='btn-info'>Listar</a> ";
+                 }
                 if (!empty($this->data['viewSitPages'])) {
+                    if( $this->data['button']['edit_sits_pages']){
                     echo "<a href='" . URLADM . "edit-sits-pages/index/" . $this->data['viewSitPages'][0]['id'] . "' class='btn-warning'>Editar</a> ";
+                    }
+                    if( $this->data['button']['delete_sits_pages']){
                     echo "<a href='" . URLADM . "delete-sits-pages/index/" . $this->data['viewSitPages'][0]['id'] . "' onclick='return confirm(\"Tem certeza que deseja excluir este registro?\")' class='btn-danger'>Apagar</a> ";
+                    }
                 }
                 ?>
             </div>
